Move interna/github and internal/temporal under activities.
