+++ server/neptune/workflows/internal/activities/deployment/info.go
@@ -1,14 +1,15 @@
-package root
+package deployment
 
 import (
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
 )
 
-type DeploymentInfo struct {
+type Info struct {
 	Version    string
 	ID         string
 	CheckRunID int64
 	Revision   string
 	Repo       github.Repo
-	Root       Root
+	Root       root.Root
 }
