+++ server/neptune/workflows/internal/activities/github.go
@@ -11,9 +11,9 @@ import (
 	"github.com/google/go-github/v45/github"
 	"github.com/hashicorp/go-getter"
 	"github.com/pkg/errors"
-	internal "github.com/runatlantis/atlantis/server/neptune/workflows/internal/github"
+	internal "github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/temporal"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/temporal"
 )
 
 type ClientContext struct {
