+++ server/neptune/workflows/internal/activities/github/link/builder.go
@@ -2,10 +2,11 @@ package link
 
 import (
 	"fmt"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/github"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
 	"net/url"
 	"path"
+
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
 )
 
 type Builder struct{}
