+++ server/neptune/workflows/internal/activities/github/link/builder_test.go
@@ -1,12 +1,13 @@
 package link_test
 
 import (
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/github"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/github/link"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
-	"github.com/stretchr/testify/assert"
 	httpurl "net/url"
 	"testing"
+
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github/link"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
+	"github.com/stretchr/testify/assert"
 )
 
 func Test_BuildDownloadLinkFromArchive(t *testing.T) {
