+++ server/neptune/workflows/internal/activities/github/markdown/templates/checkrun.tmpl
