+++ server/neptune/workflows/internal/deploy/terraform/deployment.go
@@ -2,7 +2,8 @@ package terraform
 
 import (
 	"fmt"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/github"
+
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
 
 	"github.com/google/uuid"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
