+++ server/neptune/workflows/internal/root/root.go
@@ -1,7 +1,7 @@
 package root
 
 import (
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/github"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/job"
 )
