+++ server/neptune/workflows/internal/terraform/job/cmd_step_runner_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/github"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/job"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
 	runner "github.com/runatlantis/atlantis/server/neptune/workflows/internal/terraform/job"
