+++ server/neptune/workflows/internal/terraform/request.go
@@ -1,7 +1,7 @@
 package terraform
 
 import (
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/github"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
 )
